model_app/requirements.txt
python-dotenv
psycopg2-binary
asyncpg
black
pgvector
PyPDF2
docx
python-docx
celery
openai
beautifulsoup4
kombu
amqp
langchain
langchain_community
langchain_postgres
langchain_core
langchain_experimental
httpx==0.26.0
